GMLAS: Windows compilation fixes
git-svn-id: https://svn.osgeo.org/gdal/trunk@35652 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/gmlas/ogrgmlasxlinkresolver.cpp
@@ -121,7 +121,7 @@ CPLString GMLASXLinkResolver::FetchRawContent(const CPLString& osURL,
     time_t nTimeStart = time(NULL);
     CPLHTTPResult* psResult = CPLHTTPFetch(osURL, papszOptions);
     time_t nTimeStop = time(NULL);
-    m_nGlobalResolutionTime += nTimeStop - nTimeStart;
+    m_nGlobalResolutionTime += static_cast<int>(nTimeStop - nTimeStart);
     CSLDestroy(papszOptions);
     if( psResult == NULL )
         return CPLString();
